test(plotting): move 2026-06-15 QA regression tests into existing files
Distribute the standalone test_qa_remediation_2026_06_15.py into the relevant
existing plotting test modules (exploration, diagnostics, signal, evaluation,
forecasting, model_selection, utils) and remove the standalone file.
